Wave system, state machine for build/combat phase
The UI buttons can still queue to build a tower during combat. Combat timer becomes funky if forcing a wave to spawn early. Tower button script can be greatly simplified

+++ Assets/Scripts/Enemy/WaveManager/Wave.cs
@@ -9,7 +9,7 @@ public struct EnemyGroup
 {
     public EnemyStats enemyType;
     public int amount;
-    public int spacing;
+    public float spacing;
 }
